Spatie update form

+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\AnomaliController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\KttController;
 use App\Http\Controllers\HomeController;
@@ -9,7 +10,7 @@ use App\Http\Controllers\BeritaController;
 use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\DashboardController;
 use App\Http\Controllers\PermissionController;
-
+use App\Http\Middleware\AutoPermission;
 
 Route::controller(HomeController::class)->group(function () {
     Route::get('/', 'index')->name('home');
@@ -40,23 +41,26 @@ Route::middleware(['auth'])->prefix('dashboard')->group(function () {
     Route::post('berita/create', [BeritaController::class, 'store'])->name('dashboard.berita.store');
     Route::get('berita/{slug}', [BeritaController::class, 'show'])->name('dashboard.berita.show');
 
-    Route::get('ktt', [KttController::class, 'index'])->name('dashboard.ktt.index');
+    Route::get('ktt', [KttController::class, 'index'])->middleware(AutoPermission::class . ':View Ktt')->name('dashboard.ktt.index');
     Route::post('ktt', [KttController::class, 'store'])->name('dashboard.ktt.store');
+
+    Route::get('/anomali', [AnomaliController::class, 'index'])->middleware(AutoPermission::class . ':View Anomali')->name('dashboard.anomali.index');
+    Route::get('/anomali/create', [AnomaliController::class, 'create'])->middleware(AutoPermission::class . ':View Anomali Create')->name('dashboard.anomali.create');
 
     Route::get('user', [UserController::class, 'index'])->name('dashboard.user.index');
     Route::post('user', [UserController::class, 'store'])->name('dashboard.user.store');
     Route::get('user/{id}/role', [UserController::class, 'showAssignRoleForm'])->name('dashboard.user.role.edit');
     Route::post('user/{id}/role', [UserController::class, 'updateRole'])->name('dashboard.user.role.update');
 
-    Route::get('role', [RoleController::class, 'index'])->name('role.index');
+    Route::get('role', [RoleController::class, 'index'])->name('dashboard.role.index');
     Route::post('roles', [RoleController::class, 'store']);
     Route::resource('roles', RoleController::class)->only(['edit', 'update']);
-    Route::resource('permissions', PermissionController::class)->names('permissions');
     Route::delete('roles/{role}', [RoleController::class, 'destroy']);
 
+    // Permission routes
     Route::get('permission', [PermissionController::class, 'index'])->name('permission.index');
-    Route::post('permissions', [PermissionController::class, 'store']);
-    Route::delete('permissions/{permission}', [PermissionController::class, 'destroy']);
+    Route::post('permissions', [PermissionController::class, 'store'])->name('permissions.store');
+    Route::delete('permissions/{permission}', [PermissionController::class, 'destroy'])->name('permissions.destroy');
 
     Route::get('role/{id}/permissions', [RoleController::class, 'editPermissions'])->name('role.permissions.edit');
     Route::post('role/{id}/permissions', [RoleController::class, 'updatePermissions'])->name('role.permissions.update');
